requirements.txt
torch>=1.0
jieba>=0.39
transformers>=2.0
hydra-core>=0.10
matplotlib>=3.0
scikit_learn>=0.20
